% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-age-group.R
\name{clean_age_range}
\alias{clean_age_range}
\title{Clean age based on age range}
\usage{
clean_age_range(age, id_study, age_design_min, age_design_max)
}
\arguments{
\item{age}{A vector of number.}

\item{id_study}{A vector of id_study for message purpose.}

\item{age_design_min}{A vector of number.}

\item{age_design_max}{A vector of number.}
}
\value{
A cleaned \code{age} vector accounting for \code{age_design_min} and \code{age_design_max}.
}
\description{
Clean age based on age range
}
