% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-data.R
\name{clean_categorical}
\alias{clean_categorical}
\title{Cleaning a categorical variable}
\usage{
clean_categorical(var, var_name, values, id_study)
}
\arguments{
\item{var}{a vector to be cleaned}

\item{var_name}{the name of the variable being cleaned}

\item{values}{a vector of values that are valid for thie categorical value, e.g., c(0,1)}
}
\description{
Cleaning a categorical variable
}
\seealso{
\code{\link[=clean_data]{clean_data()}}
}
