% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-age-group.R
\name{make_age_group_anthro}
\alias{make_age_group_anthro}
\title{Generate age group for anthro data}
\usage{
make_age_group_anthro(...)
}
\description{
Generate age group for anthro data
}
\seealso{
\code{\link[=make_age_group]{make_age_group()}}
}
