% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data_changes.R
\name{read_extracted_df}
\alias{read_extracted_df}
\title{Function to read previously extracted dataframe}
\usage{
read_extracted_df(
  filename,
  study_dir = NULL,
  extracted_data_dir = NULL,
  read_from_extracted_data_dir = TRUE
)
}
\arguments{
\item{filename}{name of files to be read. Use the format "Country Study_name Year_duration". For example: "USA NHANES 2005-2016"}

\item{study_dir}{location of study folder for reading extracted files}

\item{extracted_data_dir}{location of \verb{extracted survey} folder: must be specified when read_from_extracted_data_dir is TRUE}

\item{read_from_extracted_data_dir}{whether to read from \verb{extracted survey} folder (default TRUE). Set to FALSE to compare with the dataframe written in the study folder}
}
\value{
data frame of previously extracted data, or NULL if not found
}
\description{
This function reads extracted data from the specified file path.
By default it reads from the extracted survey folder. Set read_from_extracted_data_dir to FALSE to compare with the dataframe written in the study folder.
}
