% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-data-changes.R
\name{check_data_changes}
\alias{check_data_changes}
\title{Function to check for changes in data outputs}
\usage{
check_data_changes(data, filename, study_dir = NULL, extracted_data_dir = NULL)
}
\arguments{
\item{data}{data frame of new extraction data to be checked}

\item{filename}{name of files to be read. Use the format "Country Study_name Year_duration". For example: "USA NHANES 2005-2016"}

\item{study_dir}{location of study folder for reading extracted files}

\item{extracted_data_dir}{location of \verb{extracted survey} folder: must be specified when read_from_extracted_data_dir is TRUE}
}
\description{
This function compares new extraction data against previously extracted survey data to identify changes and differences.
It uses read_extracted_df to find comparison data and compare_dataframes to perform the actual comparison.
By default it reads from the extracted survey folder. Set read_from_extracted_data_dir to FALSE to compare with the dataframe written in the study folder.
}
