% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-extraction.R
\name{check_extraction}
\alias{check_extraction}
\title{Function to check extractions before saving}
\usage{
check_extraction(dataset)
}
\arguments{
\item{dataset}{data frame of extraction to be checked: can be a single study or multiple studies.}
}
\description{
This function checks formatted extractions and print out error, check and caution messages
when issues are identified. Errors need immediate attention and will stop the other checks.
Checks are usually issues that need amending and they are produced together in single run.
Cautions may not be a problem; they highlight something that require an active check and
decision to ignore.
}
\details{
Edit history
\itemize{
\item RS created May 2023
\item BZ Jan 2024: added condition for checking age range; added checks for value constraints for BP and waist/hip
\item YDDB Dec 2024: adding checks for variables that should be 0/1 - not completed yet (10/12/2024)
\item ABP March 2025: Corrected line 125 to avoid error caused by only one variable of the pairs available
\item BZ March 2025: corrected condition for pair wise comparison where integer overflow happened when multiplying two large Ns
\item BZ March 2025: added checks for empty strings or 'NA' strings in variable values; added check for unique values in metadata columns; added more comments
\item FD September 2025: added checks for hb altitude adjustment
}
}
