% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_extraction.R
\name{check_extraction}
\alias{check_extraction}
\title{Function to check extractions before saving}
\usage{
check_extraction(dataset)
}
\arguments{
\item{dataset}{data frame of extraction to be checked: can be a single study or multiple studies.}
}
\description{
This function checks formatted extractions and print out error, check and caution messages
when issues are identified. Errors need immediate attention and will stop the other checks.
Checks are usually issues that need amending and they are produced together in single run.
Cautions may not be a problem; they highlight something that require an active check and
decision to ignore.
}
