% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-data.R
\name{clean_continuous}
\alias{clean_continuous}
\title{Cleaning a continuous variable}
\usage{
clean_continuous(var, var_name, minv, maxv, id_study)
}
\arguments{
\item{var}{a vector to be cleaned}

\item{var_name}{the name of the variable being cleaned}

\item{minv}{minimum plausible value}

\item{maxv}{maximum plausible value}
}
\description{
Cleaning a continuous variable
}
\examples{
fgl_clean <- clean_continuous(data[[variable]], variable, 2, 30)
}
\seealso{
\code{\link[=clean_data]{clean_data()}}
}
