% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-age-group.R
\name{make_age_group}
\alias{make_age_group}
\title{Generate age group for each individual accounting for the age range of study}
\usage{
make_age_group(age, age_design_min, age_design_max, anthro = FALSE)
}
\arguments{
\item{age}{A vector of number.}

\item{age_design_min}{A vector of number.}

\item{age_design_max}{A vector of number.}
}
\value{
The 10-year age group and corresponding mean age of \code{age} accounting for \code{age_design_min} and \code{age_design_max}.
}
\description{
Generate age group for each individual accounting for the age range of study
}
\examples{
make_age_group(29, 18, 200)
make_age_group(c(29,32,50,80,12,18,80), c(18,18,55,55,10,10,18), c(200,35,80,80,200,200,200))
make_age_group(c(29,32,50,80,12,18,80), c(18,18,55,55,10,10,18), c(200,35,80,80,200,200,200), anthro = TRUE)
make_age_group_anthro(c(29,32,50,80,12,18,80), c(18,18,55,55,10,10,18), c(200,35,80,80,200,200,200))
}
\seealso{
\code{\link[=make_age_group_anthro]{make_age_group_anthro()}}

data$age_min <- with(data, ifelse(sex=="male", age_min_anthro_M, age_min_anthro_F)) # use appropriate variables depending on purpose
data$age_max <- with(data, ifelse(sex=="male", age_max_anthro_M, age_max_anthro_F))
ages <- make_age_groups(data$age, data$age_min, data$age_max)
data$age_mean <- ages$age_mean
data$age_group <- ages$age_group
}
