% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data_changes.R
\name{read_extracted_df}
\alias{read_extracted_df}
\title{Function to read previously extracted dataframe}
\usage{
read_extracted_df(filename, target_dir = NULL)
}
\arguments{
\item{filename}{name of file to be read without suffix}

\item{target_dir}{location of study folder for reading extracted files; default is the current study folder}
}
\value{
data frame of previously extracted data, or NULL if not found
}
\description{
This function reads extracted data from the specified file path.
}
