% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_followups.R
\name{read_followups_tracker}
\alias{read_followups_tracker}
\title{Function to read the follow-up tracker}
\usage{
read_followups_tracker(tracker = NULL)
}
\arguments{
\item{tracker}{data frame of the follow-up tracker to be used; default is the tracker packaged with \code{ncdrisc}}
}
\value{
data frame of the follow-up tracker, with the columns \code{id_study}, \code{year}, \code{fu} and \code{cohort_group_id}
}
\description{
This function returns the "Follow-ups tracker" with standardised column types.
By default the version packaged with \code{ncdrisc} is used; a data frame can be supplied
to use a more recent version of the tracker.
}
