% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_extraction.R
\name{save_extraction}
\alias{save_extraction}
\title{Function to save extraction}
\usage{
save_extraction(
  data,
  filename,
  extracted_data_dir =
    "S:/Projects/HeightProject/Original dataset/Data/Surveys/Extracted Survey/"
)
}
\arguments{
\item{data}{data frame to be saved}

\item{filename}{name of files to be saved. Use the format "Country Study_name Year_duration". For example:
"USA NHANES 2005-2016"}

\item{extracted_data_dir}{location of extracted survey folder}
}
\description{
This function adds user information to extracted data, save extraction logs and save formatted extractions.
It has to be run from the study folder.
}
